+++ Adapter/ShopwareAdapter/RequestGenerator/Product/ProductRequestGenerator.php
@@ -164,6 +164,18 @@ class ProductRequestGenerator implements ProductRequestGeneratorInterface
      */
     private function addShippingProfilesAsAttributes(Product $product)
     {
+        $allProfileIdentities = $this->identityService->findBy([
+            'objectType' => ShippingProfile::TYPE,
+            'adapterName' => ShopwareAdapter::NAME,
+        ]);
+
+        $shippingAttributes = [];
+        foreach ($allProfileIdentities as $identity) {
+            $shippingAttributes['shippingProfile' . $identity->getAdapterIdentifier()] = '';
+        }
+
+        $attributes = $product->getAttributes();
+
         foreach ($product->getShippingProfileIdentifiers() as $identifier) {
             $profileIdentity = $this->identityService->findOneBy([
                 'objectIdentifier' => $identifier,
@@ -175,8 +187,6 @@ class ProductRequestGenerator implements ProductRequestGeneratorInterface
                 continue;
             }
 
-            $attributes = $product->getAttributes();
-
             $existingAttributes = array_filter($attributes, function (Attribute $attribute) use ($profileIdentity) {
                 return $attribute->getKey() === 'shippingProfile' . $profileIdentity->getAdapterIdentifier();
             });
@@ -187,13 +197,17 @@ class ProductRequestGenerator implements ProductRequestGeneratorInterface
                 continue;
             }
 
-            $attributes[] = Attribute::fromArray([
-                'key'   => 'shippingProfile' . $profileIdentity->getAdapterIdentifier(),
-                'value' => $profileIdentity->getObjectIdentifier(),
-            ]);
-
-            $product->setAttributes($attributes);
+            $shippingAttributes['shippingProfile' . $profileIdentity->getAdapterIdentifier()] = $profileIdentity->getObjectIdentifier();
         }
+
+        foreach ($shippingAttributes as $key => $value) {
+            $attributes[] = Attribute::fromArray([
+                'key' => $key,
+                'value' => $value,
+            ]);
+        }
+
+        $product->setAttributes($attributes);
     }
 
     /**