+++ Adapter/ShopwareAdapter/ResponseParser/Order/OrderResponseParser.php
@@ -112,7 +112,7 @@ class OrderResponseParser implements OrderResponseParserInterface
 
         $orderItems = array_filter(array_map(function (array $orderItem) use ($taxFree) {
             return $this->orderItemResponseParser->parse($orderItem, $taxFree);
-        }, $this->prepareOrderItems($entry['details'], $entry['net'])));
+        }, $this->prepareOrderItems($entry['details'], (bool) $entry['net'])));
 
         $orderItems[] = $this->getShippingCosts($entry, $taxFree);
 
@@ -230,7 +230,7 @@ class OrderResponseParser implements OrderResponseParserInterface
      */
     private function prepareOrderItems(array $orderItems, $isNet)
     {
-        foreach ($orderItems as $key => $orderItem) {
+        foreach ($orderItems as $key => &$orderItem) {
             if (empty($orderItem['taxId'])) {
                 if (empty($orderItem['taxRate'])) {
                     continue;