+++ Connector/TransferObject/Order/Order.php
@@ -20,6 +20,7 @@ class Order extends AbstractTransferObject implements AttributableInterface
     const TYPE = 'Order';
 
     const TYPE_ORDER = 1;
+
     const TYPE_OFFER = 2;
 
     /**