+++ Connector/TransferObject/Product/LinkedProduct/LinkedProduct.php
@@ -9,9 +9,9 @@ use PlentyConnector\Connector\ValueObject\AbstractValueObject;
  */
 class LinkedProduct extends AbstractValueObject
 {
-    const TYPE_ACCESSORY   = 'accessory';
-    const TYPE_REPLACEMENT = 'replacement';
-    const TYPE_SIMILAR     = 'similar';
+    const TYPE_ACCESSORY = 1;
+    const TYPE_REPLACEMENT = 2;
+    const TYPE_SIMILAR = 3;
 
     /**
      * @var string