+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 - prepareOrderItems validation fix 
+- import alternate media text as name if is set
 
 ### Changed