+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - prepareOrderItems validation fix 
 - only transfer variations if has valid clientid
+- import alternate media text as name if is set
 
 ### Changed