+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - prepareOrderItems validation fix 
 - only transfer variations with valid clientid
 - import alternate media text as name if is set
+- import correct main translation
 
 ### Changed