+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - only transfer variations with valid clientid
 - import alternate media text as name if is set
 - import correct main translation
+- use correct item tax when transferring order to plenty
 
 ### Changed