+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 - changed the sequence in which the definitions are processed, orders and payments are now fetched first.
+- reset unused Shopware attributes for shipping profiles
 
 ## [4.2.1]
 ### Added